Added cleaning funktion to the folders

+++ initialize.py
@@ -1,12 +1,12 @@
 import	os
 from glob import glob
 
-class WoldCreation:
+class initialize:
 	"""docstring for Itinialise"""
 
 	Folders = {
-		'Data': ['PalmSense', 'Humidity', 'Oxygen', 'OxygenBoard'],
-		'Result': ['Response', 'RawData', 'Oxygen']
+		'Data': ['PalmSense', 'Humidity', 'Oxygen', 'OxygenBoard'], #Data Folders!
+		'Result': ['Response', 'RawData', 'TimeOxygen']
 		}
 
 	# def __init__(self, arg):
@@ -26,17 +26,20 @@ class WoldCreation:
 
 
 	def cleaner(self,i):
-		for x in glob(i+'/*'):
-			try:
-				os.remove(x)
-			except FileNotFoundError as e:
-				print(e, "\n Sorry, i can't clean files! Probably folder already empty!\n")
-			except PermissionError as e:
-				print(e, "\n Sorry, i can't clean files! Probably file in use, or opened!\n")
-			except Exception as e: 
-				print(e)
-			else:
-				print('File:' + x + 'removed!')
+		if not glob(i+'/*'):
+			print('No files in derictory: ' + i)
+		else:
+			for x in glob(i+'/*'):
+				try:
+					os.remove(x)
+				except FileNotFoundError as e:
+					print(e, "\n Sorry, i can't clean files! Probably folder already empty!\n")
+				except PermissionError as e:
+					print(e, "\n Sorry, i can't clean files! Probably file in use, or opened!\n")
+				except Exception as e: 
+					print(e)
+				else:
+					print('File:' + x + ' was removed!')
 
 	def create_folders(self):
 		#Create Subfolders in Data
@@ -45,9 +48,30 @@ class WoldCreation:
 			for x in Lvl2:
 				self.maker(Lvl1+'/'+x)
 
-	def clean_folders(self):
+	def clean_all_folders(self):
 		for Lvl1, Lvl2  in self.Folders.items(): 
 			for x in Lvl2:
 				self.cleaner(Lvl1+'/'+x)
+		print('Cleaning done!')
+
+
+	def clean_data_folders(self):
+		name = 'Data'
+		for Folder in self.Folders[name]:
+			# print(Folder) 
+			self.cleaner(name+'/'+Folder)
+		print('Cleaning data folders done!')
+
+	def clean_specific(self, GoalFolder):
+		# print(os.listdir('./'))
+		for root, dirs, files in os.walk('./'):
+			for names in dirs:
+				if names == GoalFolder:
+					# self.cleaner(name+'/'+Folder)
+					self.cleaner(root+'/'+names)
+				else:
+					pass
+
+
